fixed the incomplete index
`doseq` implementation used to return an incomplete index. The cause for
this behavior is the incorrect usage of `transient` data structures. One
should use the updated value returned by the modifier functions (`conj!`,
`assoc!`, etc) instead of the initial one, due to clojure's
implementation of transients. That required me to convert `doseq` back
to reduce and refactor some functions for better readability.